shared/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "etl-carriers"
version = "1.0.0"
description = "Shared ETL core for BetterWase carrier data pipeline"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-storage>=2.0",
    "google-cloud-bigquery>=3.0",
    "google-cloud-pubsub>=2.0",
    "pandas>=2.0",
    "openpyxl>=3.0",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["../../tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
